+++ src/templateList.tsx
@@ -5,6 +5,7 @@ import { requestAPI } from './handler';
 
 interface IProperties {
   onSelect: (selected: ISelectedTemplate) => void;
+  onContextMenu?: (selected: ISelectedTemplate) => void;
   name?: string;
   path?: string;
   isRoot?: boolean;
@@ -36,12 +37,16 @@ class TemplateFolder extends React.Component<IProperties, IState> {
             path={item.path}
             isRoot={false}
             onSelect={this.props.onSelect}
+            onContextMenu={this.props.onContextMenu}
           />
         );
       } else if (item.type === 'notebook') {
         return (
           <li className="template-item template-notebook">
-            <a onClick={() => this.props.onSelect(item)}>
+            <a
+              onClick={() => this.props.onSelect(item)}
+              onContextMenu={() => this.props.onContextMenu(item)}
+            >
               <div>
                 <span className="jp-multicontents-templates-item-icon"></span>
                 <span>{item.name}</span>
@@ -127,23 +132,30 @@ class TemplateFolder extends React.Component<IProperties, IState> {
 
 class TemplateListWidget extends ReactWidget {
   onSelect: (selected: ISelectedTemplate) => void;
+  onContextMenu: (selected: ISelectedTemplate) => void;
 
-  constructor(onSelect: (selected: ISelectedTemplate) => void) {
+  constructor(
+    onSelect: (selected: ISelectedTemplate) => void,
+    onContextMenu?: (selected: ISelectedTemplate) => void
+  ) {
     super();
     this.addClass('jp-ReactWidget');
     this.title.iconClass = 'jp-multicontents-templates-icon';
     this.title.caption = 'Templates';
     this.title.closable = true;
     this.onSelect = onSelect;
+    this.onContextMenu = onContextMenu;
   }
 
   render(): JSX.Element {
     return (
       <div className="multicontents-templates">
         <h1> Templates </h1>
-        <div className="multicontents-templates-list">
-          <TemplateFolder isRoot={true} onSelect={this.onSelect} />
-        </div>
+        <TemplateFolder
+          isRoot={true}
+          onSelect={this.onSelect}
+          onContextMenu={this.onContextMenu}
+        />
       </div>
     );
   }