+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-# v 0.2.0 
-- Allow publish templates from filebrowser 
+# v 0.3.0
+- Allow share the template with URL, and people with the same URL with open the preview window on their server
+# v 0.2.0
+- Allow publish templates from filebrowser
 - Added loading icon when loading template list
 - Fixed a bug that failed to preview template when file name contains html special char